contrib/database/sql: rewrite
