+++ contrib/database/sql/internal/postgres.go
@@ -1,7 +1,16 @@
-package pq
+// Copied from package github.com/lib/pq:
+//
+// parseOpts: https://github.com/lib/pq/blob/61fe37aa2ee24fabcdbe5c4ac1d4ac566f88f345/conn.go
+// parseURL: https://github.com/lib/pq/blob/50761b0867bd1d9d069276790bcd4a3bccf2324a/url.go
+
+package internal
 
 import (
 	"fmt"
+	"net"
+	nurl "net/url"
+	"sort"
+	"strings"
 	"unicode"
 )
 
@@ -39,9 +48,9 @@ func (s *scanner) SkipSpaces() (rune, bool) {
 	return r, ok
 }
 
-// ParseOpts parses the options from name and adds them to the values.
+// parseOpts parses the options from name and adds them to the values.
 // The parsing code is based on conninfo_parse from libpq's fe-connect.c
-func ParseOpts(name string, o values) error {
+func parseOpts(name string, o values) error {
 	s := newScanner(name)
 
 	for {
@@ -115,4 +124,71 @@ func ParseOpts(name string, o values) error {
 	}
 
 	return nil
+}
+
+// parseURL no longer needs to be used by clients of this library since supplying a URL as a
+// connection string to sql.Open() is now supported:
+//
+//    sql.Open("postgres", "postgres://bob:secret@1.2.3.4:5432/mydb?sslmode=verify-full")
+//
+// It remains exported here for backwards-compatibility.
+//
+// parseURL converts a url to a connection string for driver.Open.
+// Example:
+//
+//    "postgres://bob:secret@1.2.3.4:5432/mydb?sslmode=verify-full"
+//
+// converts to:
+//
+//    "user=bob password=secret host=1.2.3.4 port=5432 dbname=mydb sslmode=verify-full"
+//
+// A minimal example:
+//
+//    "postgres://"
+//
+// This will be blank, causing driver.Open to use all of the defaults
+func parseURL(url string) (string, error) {
+	u, err := nurl.Parse(url)
+	if err != nil {
+		return "", err
+	}
+
+	if u.Scheme != "postgres" && u.Scheme != "postgresql" {
+		return "", fmt.Errorf("invalid connection protocol: %s", u.Scheme)
+	}
+
+	var kvs []string
+	escaper := strings.NewReplacer(` `, `\ `, `'`, `\'`, `\`, `\\`)
+	accrue := func(k, v string) {
+		if v != "" {
+			kvs = append(kvs, k+"="+escaper.Replace(v))
+		}
+	}
+
+	if u.User != nil {
+		v := u.User.Username()
+		accrue("user", v)
+
+		v, _ = u.User.Password()
+		accrue("password", v)
+	}
+
+	if host, port, err := net.SplitHostPort(u.Host); err != nil {
+		accrue("host", u.Host)
+	} else {
+		accrue("host", host)
+		accrue("port", port)
+	}
+
+	if u.Path != "" {
+		accrue("dbname", u.Path[1:])
+	}
+
+	q := u.Query()
+	for k := range q {
+		accrue(k, q.Get(k))
+	}
+
+	sort.Strings(kvs) // Makes testing easier (not a performance concern)
+	return strings.Join(kvs, " "), nil
 }
