--- contrib/database/sql/parsedsn/mysql/mysql.go
@@ -1,3 +1,0 @@
-// Package mysql is the minimal fork of go-sql-driver/mysql so we can use their code
-// to parse the mysql DSNs
-package mysql
