--- contrib/database/sql/parsedsn/pq/pq.go
@@ -1,2 +1,0 @@
-// Package pq is the minimal fork of lib/pq so we can use their code to parse the postgres DSNs
-package pq
