--- contrib/database/sql/sqlutils/sqlutils.go
@@ -1,2 +1,0 @@
-// Package sqlutils share some utils functions for sql packages
-package sqlutils
